Reorganized the files for a better overview
